requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.8
django-cloudinary-storage==0.3.0
idna==3.11
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn
whitenoise
django-ratelimit